requirements.txt
flask
flask-sqlalchemy
flask-login
flask-wtf
flask-admin
email-validator
gunicorn
flask-dance
flask-login
oauthlib
pyjwt
sqlalchemy
wtforms
